+++ src/AutoV/Spec/projection/rules.cpp
@@ -1385,8 +1385,8 @@ static vector<int> PRIM_NUMS = get_prime();
 
 SpecNode* try_divide_const_factor(SpecNode *expr, int factor) {
     if(auto m = instance_of(expr, IntConst)) {
-        if(std::get<long long>(m->value) != 0 && std::get<long long>(m->value) % factor == 0) {
-            return new IntConst(std::get<long long>(m->value) / factor);
+        if(std::get<unsigned long>(m->value) != 0 && std::get<unsigned long>(m->value) % factor == 0) {
+            return new IntConst(std::get<unsigned long>(m->value) / factor);
         }
     } else if(auto m = instance_of(expr, Expr)) {
         if(holds_alternative<Expr::binops>(m->op)) {
@@ -1729,52 +1729,52 @@ rule_ret_t rule_simplify_expr(Project *proj, SpecNode *spec) {
 
                     if(ops == op::ADD) {
                         expr_is_changed = true;
-                        expr.reset(new IntConst(std::get<long long>(elem0->value) + std::get<long long>(elem1->value)));
+                        expr.reset(new IntConst(std::get<unsigned long>(elem0->value) + std::get<unsigned long>(elem1->value)));
                     } else if (ops == op::MINUS && m->elems->size() == 1) {
                         expr_is_changed = true;
-                        expr.reset(new IntConst(-std::get<long long>(elem0->value)));
+                        expr.reset(new IntConst(-std::get<unsigned long>(elem0->value)));
                     } else if (ops == op::MINUS && m->elems->size() == 2) {
                         expr_is_changed = true;
-                        expr.reset(new IntConst(std::get<long long>(elem0->value) - std::get<long long>(elem1->value)));
+                        expr.reset(new IntConst(std::get<unsigned long>(elem0->value) - std::get<unsigned long>(elem1->value)));
                     } else if (ops == op::MULT) {
                         expr_is_changed = true;
-                        expr.reset(new IntConst(std::get<long long>(elem0->value) * std::get<long long>(elem1->value)));
+                        expr.reset(new IntConst(std::get<unsigned long>(elem0->value) * std::get<unsigned long>(elem1->value)));
                     } else if(ops == op::DIV) {
                         expr_is_changed = true;
-                        expr.reset(new IntConst(std::get<long long>(elem0->value) / std::get<long long>(elem1->value)));
+                        expr.reset(new IntConst(std::get<unsigned long>(elem0->value) / std::get<unsigned long>(elem1->value)));
                     } else if(ops == op::MOD) {
                         expr_is_changed = true;
-                        expr.reset(new IntConst(std::get<long long>(elem0->value) + std::get<long long>(elem1->value)));
+                        expr.reset(new IntConst(std::get<unsigned long>(elem0->value) + std::get<unsigned long>(elem1->value)));
                     } else if(ops == op::BITAND) {
                         expr_is_changed = true;
-                        expr.reset(new IntConst(std::get<long long>(elem0->value) & std::get<long long>(elem1->value)));
+                        expr.reset(new IntConst(std::get<unsigned long>(elem0->value) & std::get<unsigned long>(elem1->value)));
                     } else if(ops == op::BITOR) {
                         expr_is_changed = true;
-                        expr.reset(new IntConst(std::get<long long>(elem0->value) | std::get<long long>(elem1->value)));
+                        expr.reset(new IntConst(std::get<unsigned long>(elem0->value) | std::get<unsigned long>(elem1->value)));
                     } else if(ops == op::LSHIFT) {
                         expr_is_changed = true;
-                        expr.reset(new IntConst(std::get<long long>(elem0->value) << std::get<long long>(elem1->value)));
+                        expr.reset(new IntConst(std::get<unsigned long>(elem0->value) << std::get<unsigned long>(elem1->value)));
                     } else if(ops == op::RSHIFT) {
                         expr_is_changed = true;
-                        expr.reset(new IntConst(std::get<long long>(elem0->value) >> std::get<long long>(elem1->value)));
+                        expr.reset(new IntConst(std::get<unsigned long>(elem0->value) >> std::get<unsigned long>(elem1->value)));
                     } else if(ops == op::BEQ) {
                         expr_is_changed = true;
-                        expr.reset(new BoolConst(std::get<long long>(elem0->value) == std::get<long long>(elem1->value)));
+                        expr.reset(new BoolConst(std::get<unsigned long>(elem0->value) == std::get<unsigned long>(elem1->value)));
                     } else if(ops == op::BNE) {
                         expr_is_changed = true;
-                        expr.reset(new BoolConst(std::get<long long>(elem0->value) != std::get<long long>(elem1->value)));
+                        expr.reset(new BoolConst(std::get<unsigned long>(elem0->value) != std::get<unsigned long>(elem1->value)));
                     } else if(ops == op::BGT) {
                         expr_is_changed = true;
-                        expr.reset(new BoolConst(std::get<long long>(elem0->value) > std::get<long long>(elem1->value)));
+                        expr.reset(new BoolConst(std::get<unsigned long>(elem0->value) > std::get<unsigned long>(elem1->value)));
                     } else if(ops == op::BLT) {
                         expr_is_changed = true;
-                        expr.reset(new BoolConst(std::get<long long>(elem0->value) < std::get<long long>(elem1->value)));
+                        expr.reset(new BoolConst(std::get<unsigned long>(elem0->value) < std::get<unsigned long>(elem1->value)));
                     } else if(ops == op::BGE) {
                         expr_is_changed = true;
-                        expr.reset(new BoolConst(std::get<long long>(elem0->value) >= std::get<long long>(elem1->value)));
+                        expr.reset(new BoolConst(std::get<unsigned long>(elem0->value) >= std::get<unsigned long>(elem1->value)));
                     } else if(ops == op::BLE) {
                         expr_is_changed = true;
-                        expr.reset(new BoolConst(std::get<long long>(elem0->value) <= std::get<long long>(elem1->value)));
+                        expr.reset(new BoolConst(std::get<unsigned long>(elem0->value) <= std::get<unsigned long>(elem1->value)));
                     }
 
                     expr->_str = "";
